requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
astroid==3.3.11
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromedriver==2.24.1
cryptography==45.0.7
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
invoke==2.2.0
isort==6.0.1
jsonschema==3.2.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==25.0
paramiko==4.0.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
pylint==3.3.8
pymongo==4.14.1
PyNaCl==1.5.0
pyrsistent==0.20.0
PySocks==1.7.1
pytest==8.4.2
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.32.5
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
texttable==1.7.0
tomli==2.2.1
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
